refactor: move utils/types.ts to types.ts (VF-1621)
