+++ src/index.ts
@@ -1,4 +1,4 @@
 export * from './constants';
 export * as Crypto from './crypto';
+export * from './types';
 export * from './utils';
-export * from './utils/types';
